OutputView.java 생성
