+++ src/main/java/blackjack/DecideResult.java
@@ -1,5 +1,7 @@
 package blackjack;
 
+import blackjack.view.OutView;
+
 import java.util.List;
 import java.util.Map;
 
@@ -26,17 +28,9 @@ public class DecideResult {
                     entry.getValue().WinOrLoss(0);
                 }
             } else {
-                win +=1;
+                win += 1;
             }
         }
-
-        System.out.println("## 최종 승패");
-        System.out.println("딜러: " + win + "승 " + loss + '패');
-
-        for (Map.Entry<String, Participant.ParticipantState> entry : participantStates.entrySet()) {
-            String name = entry.getValue().printname();
-            System.out.println(name + ": " + entry.getValue().printwinorloss());
-
-        }
+        OutView.printFinalResults(win, loss, participantStates);
     }
 }
