--- src/main/java/blackjack/OutView.java
@@ -1,4 +1,0 @@
-package blackjack;
-
-public class OutView {
-}
